+++ src/main/java/org/laga/moneygestor/db/repository/TransactionRepository.java
@@ -12,9 +12,9 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface TransactionRepository extends JpaRepository<TransactionDb, Integer> {
+public interface TransactionRepository extends JpaRepository<TransactionDb, Long> {
     @Transactional
     @Modifying
     @Query(value = "UPDATE transaction SET Description = :description, Value = :value, Date = :date, Wallet = :wallet, type = :type WHERE Id = :id AND User = :userId", nativeQuery = true)
-    int editWallet(@Param("id") Integer transactionId, @Param("description") String description, @Param("value") BigDecimal value, @Param("date") String date, @Param("wallet") Integer walletId, @Param("type") Integer typeId, @Param("userId") Integer userId);
+    int editWallet(@Param("id") Long transactionId, @Param("description") String description, @Param("value") BigDecimal value, @Param("date") String date, @Param("wallet") Integer walletId, @Param("type") Integer typeId, @Param("userId") Integer userId);
 }
